Added liked events to user view. Refactored ProfileController methods as UsersController methods.
